Add AssemblerExitCode and ObjSectionType enums to types.hpp
